requirements.txt
openpyxl
pyyaml
python-dateutil
tkcalendar
holidays
pytest
python-docx
ruff